+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/search_engine/selection/__init__.py
@@ -1,14 +1,18 @@
 """Index selection strategies package."""
 
 from .base import BaseIndexSelector
-from .cache_manager import IndexAliasLoader, IndexCacheManager
+from .cache_manager import IndexAliasLoader
 from .factory import IndexSelectorFactory
-from .selectors import DatetimeBasedIndexSelector, UnfilteredIndexSelector
+from .selectors import (
+    DatetimeBasedIndexSelector,
+    SyncDatetimeBasedIndexSelector,
+    UnfilteredIndexSelector,
+)
 
 __all__ = [
-    "IndexCacheManager",
     "IndexAliasLoader",
     "DatetimeBasedIndexSelector",
+    "SyncDatetimeBasedIndexSelector",
     "UnfilteredIndexSelector",
     "IndexSelectorFactory",
     "BaseIndexSelector",