+++ stac_fastapi/tests/api/test_api_datetime_filtering.py
@@ -349,9 +349,7 @@ async def test_bulk_create_items_with_early_date_in_second_batch_for_datetime_in
 
     indices = await txn_client.database.client.indices.get_alias(index="*")
     expected_aliases = [
-        "items_start_datetime_test-collection_2008-01-15-2010-02-09",
-        "items_start_datetime_test-collection_2020-02-09",
-        "items_start_datetime_test-collection_2010-02-10-2020-02-08",
+        "items_start_datetime_test-collection_2008-01-15",
     ]
 
     all_aliases = set()